Updates for scheduling algorithm files
